+++ tests/testthat.R
@@ -1,6 +1,8 @@
 library(testthat)
 library(beautier)
 
+
+unlink(get_beautier_folder(), recursive = TRUE)
 check_empty_beautier_folder()
 
 test_check("beautier")