+++ tests/testthat/test-check_screenlog.R
@@ -65,6 +65,14 @@ test_that("abuse, by wrong values", {
   expect_error(
     check_screenlog(
       create_screenlog(
+        filename = "filename with spaces.csv"
+      )
+    ),
+    "filename"
+  )
+  expect_error(
+    check_screenlog(
+      create_screenlog(
         filename = 1234
       )
     ),